reorganizes data-permissions.ts to be a folder split by a file per concern, moves some logic out into other utils files where it makes sense
