+++ frontend/src/metabase/admin/permissions/selectors/data-permissions/diff.ts
@@ -1,7 +1,7 @@
 import { createSelector } from "@reduxjs/toolkit";
 import _ from "underscore";
 
-import { diffDataPermissions } from "metabase/admin/permissions/utils/graph";
+import { diffDataPermissions } from "metabase/admin/permissions/utils/graph/permissions-diff";
 import { PLUGIN_DATA_PERMISSIONS } from "metabase/plugins";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type { Group } from "metabase-types/api";
