+++ frontend/src/metabase/admin/permissions/utils/graph/data-permissions/index.ts
@@ -1,0 +1,4 @@
+export * from "./get";
+export * from "./has";
+export * from "./update";
+export * from "./utils";
