+++ frontend/src/metabase/admin/permissions/utils/graph/data-permissions/utils.ts
@@ -1,0 +1,8 @@
+import _ from "underscore";
+
+import { DataPermissionValue } from "metabase/admin/permissions/types";
+import { PLUGIN_ADVANCED_PERMISSIONS } from "metabase/plugins";
+
+export const isRestrictivePermission = (value: DataPermissionValue) =>
+  value === DataPermissionValue.NO ||
+  PLUGIN_ADVANCED_PERMISSIONS.isRestrictivePermission(value);
